server: do bidirectional exchange for connection hello
For TLS the client (the orchestrator) sends the keys to both
endpoints. For PSP the endpoints generate the key locally so
we can't use this trick.

To be able to negotiate PSP enablement in the future send conn id
in both directions. We will add flags to the same message.
Pass ID 0/0 to the worker, it doesn't actually matter.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>