server: move patbuf initialization pre-fork in prep for devmem TX
Add earlier initialization of patbuf. When later patches bring online
devmem TX and the TX bounce buffer is initialized in the session (and
subsequently shared with workers by dmabuf), we need to also initialize
patbuf inside the session instead of the worker.

Workers still see the initialized pattern in their unmodified read-access
pages, so no additional calls in the worker are needed.

Signed-off-by: Bobby Eshleman <bobbyeshleman@meta.com>