iou: add io_uring IO backend
Add a basic io_uring IO backend to kperf with basic send and recv
support.

worker_iou_init() sets the io_ops to io_uring that the worker will use.

Add a new ended flag to worker_state to mark that a test has ended. This
is needed since there may be completions left in the completion queue
that refer to an freed connection already.

Signed-off-by: David Wei <dw@davidwei.uk>